Implemented the MVVM design pattern
In this commit we can see that the functionality of our app is completely identical to that of the previous version. So what has changed? We added a new file to the project where we put all the logic of our view. This will make the project look cleaner but most importantly much more scalable.